requirements.txt
pytorch
torchvision
torchmetrics
pillow
numpy
matplotlib
pytorch-lightning
